feat: reorg code and add exit code handling logic
